+++ lib/persistent_window_manager.dart
@@ -1,2 +1,2 @@
-export 'src/pe_wi_ma_wrapper.dart' show PeWiMaWrapper;
+export 'src/run_app_persistent_window_manager.dart' show runAppPersistentWindowManager;
 export 'src/models/custom_window_options.dart' show CustomWindowOptions;